requirements.txt
appnope
astroid
asttokens
boto3
botocore
comm
confluent-kafka
coverage
debugpy
decorator
dill
executing
iniconfig
ipykernel
ipython
isort
jedi
jmespath
jupyter_client
jupyter_core
matplotlib-inline
mccabe
nest-asyncio
packaging
parso
pexpect
platformdirs
pluggy
prompt-toolkit
psutil
psycopg2-binary
ptyprocess
pure-eval
Pygments
pylint
pytest
pytest-cov
pytest-dotenv
python-dateutil
python-dotenv
pyzmq
s3transfer
six
stack-data
tomlkit
tornado
traitlets
urllib3
wcwidth
